Refactor analysis utils
