+++ e2e_sae/scripts/analysis/geometric_analysis.py
@@ -27,13 +27,13 @@ from torch import Tensor
 from tqdm import tqdm
 from wandb.apis.public import Run
 
-from e2e_sae.analysis import get_df_gpt2
 from e2e_sae.log import logger
 from e2e_sae.plotting import plot_facet
 from e2e_sae.scripts.analysis.plot_settings import (
     SIMILAR_RUN_INFO,
     STYLE_MAP,
 )
+from e2e_sae.scripts.analysis.utils import get_df_gpt2
 from e2e_sae.settings import REPO_ROOT
